% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asfr_p_scene.R
\name{asfr_p_scene}
\alias{asfr_p_scene}
\title{Obtain ASFR parameter in Different Scenarios}
\usage{
asfr_p_scene(infile, inparm_scene)
}
\arguments{
\item{inparm_scene}{}
}
\description{
Obtain ASFR parameter in Different Scenarios
}
\examples{
asfr = asfr_p_scene(infile_pop, inparm_scene)
}
