% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.pop_move_s.R
\name{f.pop_move_s}
\alias{f.pop_move_s}
\title{Title}
\usage{
f.pop_move_s(
  pop_s,
  project_year,
  age_group,
  Lx_male,
  Lx_female,
  asfr,
  srb,
  move_net
)
}
\arguments{
\item{move_net}{}
}
\description{
Title
}
