% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.pop_move_t.R
\name{f.pop_move_t}
\alias{f.pop_move_t}
\title{Fast Obtain Population Summary Data include Immigration}
\usage{
f.pop_move_t(pop_move_s, project_year, age_group)
}
\arguments{
\item{age_group}{}
}
\description{
Fast Obtain Population Summary Data include Immigration
}
\examples{
pop_move_t = f.pop_move_t(pop_move_s, project_year, age_group)
}
