% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.pop_s.R
\name{f.pop_s}
\alias{f.pop_s}
\title{Fast Obtain Population Structure}
\usage{
f.pop_s(infile_pop, project_year, age_group, Lx_male, Lx_female, asfr, srb)
}
\arguments{
\item{srb}{}
}
\description{
Stable Population exclude immigration
}
\examples{
pop_s = f.pop_s(infile_pop, project_year, age_group, Lx_male, Lx_female, asfr, srb)
}
