% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.pop_t.R
\name{f.pop_t}
\alias{f.pop_t}
\title{Fast Converting Population Structure into Summary Data}
\usage{
f.pop_t(pop_s, project_year, age_group)
}
\arguments{
\item{age_group}{}
}
\description{
Aggregated data, which displays as 5 age group data
}
\examples{
pop_t = f.pop_t(pop_s, project_year, age_group)
}
