% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.projection.R
\name{f.projection}
\alias{f.projection}
\title{Fast Projection and Output to Excel}
\usage{
f.projection(infile_pop, inparm_scene, inparm_lt, path_out_table)
}
\arguments{
\item{path_out_table}{}
}
\description{
Fast Projection and Output to Excel
}
\examples{
path_out_table = '~/Desktop/Pop_Projection/2_Output_Table'
f.projection(infile_pop, inparm_scene, inparm_lt, path_out_table)
}
