% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.projection2word.R
\name{f.projection2word}
\alias{f.projection2word}
\title{Title}
\usage{
f.projection2word(
  path_out_table,
  path_out_pic,
  path_out,
  initial_year = 2020,
  x_by = 5,
  area_name,
  pop_name
)
}
\arguments{
\item{pop_name}{}
}
\description{
Title
}
