% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infile_pop.R
\docType{data}
\name{infile_pop}
\alias{infile_pop}
\title{Data infile_pop}
\format{
An object of class \code{list} of length 2.
}
\usage{
infile_pop
}
\description{
Data infile_pop
}
\references{
\url{http://stats.gd.gov.cn/gdtjnj/index.html}
}
\author{
Yu-Teng Yan \email{yanyuteng@ruc.edu.cn}
}
\keyword{datasets}
