% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inparm_scene.R
\docType{data}
\name{inparm_scene}
\alias{inparm_scene}
\title{Data inparm_scene}
\format{
An object of class \code{list} of length 6.
}
\usage{
inparm_scene
}
\description{
Data inparm_scene
}
\references{
\url{http://stats.gd.gov.cn/gdtjnj/index.html}
}
\author{
Yu-Teng Yan \email{yanyuteng@ruc.edu.cn}
}
\keyword{datasets}
