% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_1year.R
\name{lt_1year}
\alias{lt_1year}
\title{Build a life table}
\usage{
lt_1year(age_name, nMx)
}
\arguments{
\item{age_name}{age group}

\item{nMx}{mortality}
}
\value{
a dataframe
}
\description{
Build a life table
}
