% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lx_p_list_scene.R
\name{Lx_p_list_scene}
\alias{Lx_p_list_scene}
\title{Project Mortality Corresponding to Different Scenarios}
\usage{
Lx_p_list_scene(inparm_lt, inparm_scene, sex)
}
\arguments{
\item{sex}{}
}
\description{
see Lx_p function
}
\examples{
Lx_male <- Lx_p_list_scene(inparm_lt, inparm_scene, 'male')
Lx_female <- Lx_p_list_scene(inparm_lt, inparm_scene, 'female')
}
