% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_p_scene.R
\name{param_p_scene}
\alias{param_p_scene}
\title{Obtain Parameters in Different Scenarios}
\usage{
param_p_scene(inparm_scene, type)
}
\arguments{
\item{type}{}
}
\description{
Obtain Parameters in Different Scenarios
}
\examples{
move_net = param_p_scene(inparm_scene,'move')
srb = param_p_scene(inparm_scene,'srb')
project_year = length(inparm_scene[[1]][,1])
age_group = length(infile_pop[[1]][,1])
}
