% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_structure_baseline_scene.R
\name{pop_structure_baseline_scene}
\alias{pop_structure_baseline_scene}
\title{Initial Baseline Population Structure Corresponding to Different Scenarios}
\usage{
pop_structure_baseline_scene(pop_s, infile)
}
\arguments{
\item{infile}{}
}
\description{
see pop_structure_baseline function
}
