% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_structure_cum.R
\name{pop_structure_cum}
\alias{pop_structure_cum}
\title{Summary Projection Population}
\usage{
pop_structure_cum(
  pop_s,
  project_year = 2,
  age_group = 101,
  Lx_male,
  Lx_female,
  asfr,
  srb
)
}
\arguments{
\item{srb}{}
}
\description{
Summary Projection Population
}
