% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_structure_cum_scene.R
\name{pop_structure_cum_scene}
\alias{pop_structure_cum_scene}
\title{Summary Projection Population Corresponding to Different Scenarios}
\usage{
pop_structure_cum_scene(
  pop_s,
  project_year = 2,
  age_group = 101,
  Lx_male,
  Lx_female,
  asfr,
  srb
)
}
\arguments{
\item{srb}{}
}
\description{
see pop_structure_cum function
}
