% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_structure_immigration_cum.R
\name{pop_structure_immigration_cum}
\alias{pop_structure_immigration_cum}
\title{pop_structure_immigration_cum}
\usage{
pop_structure_immigration_cum(
  pop_s,
  move_s,
  project_year = 2,
  age_group = 101,
  Lx_male,
  Lx_female,
  asfr,
  srb
)
}
\arguments{
\item{srb}{}
}
\description{
pop_structure_immigration_cum
}
