% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_structure_immigration_cum.u.R
\name{pop_structure_immigration_cum.u}
\alias{pop_structure_immigration_cum.u}
\title{pop_structure_immigration_cum.u}
\usage{
pop_structure_immigration_cum.u(
  pop_s,
  move_s,
  urban,
  project_year = 2,
  age_group = 101,
  Lx_male,
  Lx_female,
  asfr,
  srb
)
}
\arguments{
\item{srb}{}
}
\description{
Immigrants are transferred according to the population of the current year, not according to the population of the previous year.
}
