% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_structure_immigration_cum_scene.R
\name{pop_structure_immigration_cum_scene}
\alias{pop_structure_immigration_cum_scene}
\title{pop_structure_immigration_cum_scene}
\usage{
pop_structure_immigration_cum_scene(
  pop_s,
  move_s,
  project_year = 2,
  age_group = 101,
  Lx_male,
  Lx_female,
  asfr,
  srb
)
}
\arguments{
\item{srb}{}
}
\description{
pop_structure_immigration_cum_scene
}
