% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_structure_immigration_scene.R
\name{pop_structure_immigration_scene}
\alias{pop_structure_immigration_scene}
\title{Title}
\usage{
pop_structure_immigration_scene(
  pop_s,
  move_net,
  project_year = 1,
  age_group = 101
)
}
\arguments{
\item{age_group}{}
}
\description{
Title
}
