% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_structure_initial_scene.R
\name{pop_structure_initial_scene}
\alias{pop_structure_initial_scene}
\title{pop_structure_initial_scene}
\usage{
pop_structure_initial_scene(project_year, age_group)
}
\arguments{
\item{age_group}{}
}
\description{
pop_structure_initial_scene
}
