% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_total_cum_scene.R
\name{pop_total_cum_scene}
\alias{pop_total_cum_scene}
\title{Title}
\usage{
pop_total_cum_scene(
  pop_t,
  pop_s,
  project_year = 1,
  age_group = 101,
  sex = "all"
)
}
\arguments{
\item{sex}{}
}
\description{
Title
}
