% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_total_initial_scene.R
\name{pop_total_initial_scene}
\alias{pop_total_initial_scene}
\title{pop_total_initial_scene}
\usage{
pop_total_initial_scene(project_year)
}
\arguments{
\item{project_year}{}
}
\description{
pop_total_initial_scene
}
