% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_female_15t49.R
\name{project_female_15t49}
\alias{project_female_15t49}
\title{project_female_15t49}
\usage{
project_female_15t49(initial_year = 2020, project_year = 1, dt)
}
\arguments{
\item{dt}{}
}
\description{
project_female_15t49
}
