% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_pic.R
\name{project_pic}
\alias{project_pic}
\title{Plot Projection Population}
\usage{
project_pic(dt, initial_year, project_year, x_by, path_out_pic)
}
\arguments{
\item{path_out_pic}{}
}
\description{
Plot Projection Population
}
