% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_table.R
\name{project_table}
\alias{project_table}
\title{Table Projection Population}
\usage{
project_table(initial_year = 2020, project_year = 1, dt)
}
\arguments{
\item{dt}{}
}
\description{
Table Projection Population
}
