% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_word_2pop.R
\name{project_word_2pop}
\alias{project_word_2pop}
\title{Title}
\usage{
project_word_2pop(
  dt1,
  dt2,
  path_out_pic,
  path_out,
  initial_year,
  project_year,
  area_name,
  pop_name
)
}
\arguments{
\item{pop_name}{}
}
\description{
Title
}
