% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_param.R
\name{read_param}
\alias{read_param}
\title{Obtain Projection Parameters from Excel}
\usage{
read_param(path)
}
\arguments{
\item{path}{}
}
\value{
a list
}
\description{
Obtain Projection Parameters from Excel
}
\examples{
inparm_scene = read_param('~/desktop/Pop_Projection/1_Input/Data_Parameter.xlsx')
}
