% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pop.R
\name{read_pop}
\alias{read_pop}
\title{Obtain Baseline Population from Excel}
\usage{
read_pop(path)
}
\arguments{
\item{path}{}
}
\value{
a list
}
\description{
Obtain Baseline Population from Excel
}
\examples{
infile_pop = read_pop('~/desktop/Pop_Projection/1_Input/Data_Population.xlsx')
}
