% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pop.s.R
\name{read_pop.s}
\alias{read_pop.s}
\title{Obtain Projection Population Structure from Excel}
\usage{
read_pop.s(path_out_table, project_year)
}
\arguments{
\item{project_year}{}
}
\description{
Obtain Projection Population Structure from Excel
}
\examples{
project_p_s = read_pop.s(path_out_table,project_year)
}
