% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pop.t.R
\name{read_pop.t}
\alias{read_pop.t}
\title{Obtain Projection Population Summary Data from Excel}
\usage{
read_pop.t(path_out_table)
}
\arguments{
\item{path_out_table}{}
}
\description{
Obtain Projection Population Summary Data from Excel
}
\examples{
path_out_table = '~/Desktop/0_Projection/Pop_Projection/2_Output_Table'
project_p = read_pop.t(path_out_table)
}
