% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pop_total_scene.u.R
\name{write_pop_total_scene.u}
\alias{write_pop_total_scene.u}
\title{Title}
\usage{
write_pop_total_scene.u(pop_t, path, move_include = FALSE)
}
\arguments{
\item{move_include}{}
}
\description{
Title
}
