% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoy_growth.R
\name{yoy_growth}
\alias{yoy_growth}
\title{Urbanization Rate trans to Year over Year Growth}
\usage{
yoy_growth(dt, type = 1)
}
\arguments{
\item{type}{}
}
\description{
Urbanization rate needs to be converted into an increase rate or decrease rate.
}
