% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoy_growth_scene.R
\name{yoy_growth_scene}
\alias{yoy_growth_scene}
\title{Urbanization Rate trans to Year over Year Growth in Different Scenarios}
\usage{
yoy_growth_scene(dt)
}
\arguments{
\item{dt}{}
}
\description{
Urbanization Rate trans to Year over Year Growth in Different Scenarios
}
