% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.projection2word.R
\name{f.projection2word}
\alias{f.projection2word}
\title{Fast Projection and Output to Word}
\usage{
f.projection2word(
  path_out_table,
  path_out_pic,
  path_out,
  initial_year = 2020,
  x_by = 5,
  area_name,
  pop_name
)
}
\arguments{
\item{path_out_table}{the excel output path}

\item{path_out_pic}{the picture output path}

\item{path_out}{the word output path}

\item{initial_year}{the initial projection year}

\item{x_by}{the plot year intervel}

\item{area_name}{the projection area/district}

\item{pop_name}{the projection 2 population names}
}
\description{
Based on the f.projection function, before run this function, the user should
run f.projection to get the output in xlsx format
}
\examples{
path_out_pic = c('~/desktop/Pop_Projection/3_Output_Pic_pop1','
~/desktop/Pop_Projection/3_Output_Pic_pop2')
path_out = c('~/desktop/Pop_Projection')
area_name = 'random county'
pop_name = c('pop1','pop2')
f.projection2word(path_out_table, path_out_pic, path_out,
initial_year = 2020, x_by = 5, area_name, pop_name)
}
