% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_structure_immigration.R
\name{pop_structure_immigration}
\alias{pop_structure_immigration}
\title{pop_structure_immigration}
\usage{
pop_structure_immigration(pop_s, move_net, age_group = 101, project_year = 1)
}
\arguments{
\item{project_year}{}
}
\description{
Immigrants are distributed by age according to the current year's population structure.
}
