% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_structure_immigration.u.R
\name{pop_structure_immigration.u}
\alias{pop_structure_immigration.u}
\title{Title}
\usage{
pop_structure_immigration.u(
  pop_s,
  move_net,
  urban,
  age_group = 101,
  project_year = 1
)
}
\arguments{
\item{project_year}{}
}
\description{
Title
}
