% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pop_total_scene.R
\name{write_pop_total_scene}
\alias{write_pop_total_scene}
\title{write_pop_total_scene}
\usage{
write_pop_total_scene(pop_t, path, move_include = FALSE)
}
\arguments{
\item{move_include}{}
}
\description{
write_pop_total_scene
}
